denoising-autoencoder-image-classifier/noise-image-generator/requirements.txt
opencv-python
scikit-image
gradio
numpy
pillow